research/archives/indomitable-2026-02-14-backup/requirements.txt
# SEC Filings Analysis Pipeline - Python Dependencies
# Install: pip install -r requirements.txt

psycopg2-binary==2.9.11
requests==2.32.5
beautifulsoup4==4.14.3
tiktoken==0.12.0
anthropic==0.79.0
